Update dependencies and improve scheduled function handling
- Upgrade @supabase/supabase-js to version 2.50.0
- Upgrade vitest to version 1.6.1
- Change parameter name in scheduled function from ScheduledEvent to ScheduledController
- Update compression handling for archived data with proper content encoding
- Remove obsolete archive tests
