+++ track-aggregator/package.json
@@ -7,15 +7,15 @@
     "dev": "wrangler dev",
     "start": "wrangler dev",
     "cf-typegen": "wrangler types",
-    "test": "vitest"
+    "test": "vitest --run"
   },
   "dependencies": {
-    "@supabase/supabase-js": "^2.44.4"
+    "@supabase/supabase-js": "^2.50.0"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20240722.0",
     "typescript": "^5.5.2",
-    "wrangler": "^4.13.2",
-    "vitest": "^1.4.0"
+    "vitest": "^1.6.1",
+    "wrangler": "^4.13.2"
   }
 }
